server/requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
pycparser==2.19
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
